Identify lines referring to included files.
Now the API returns `source: {line: num, file: x}` where x
is null for the user-supplied file and a path otherwise.

Hopefully fixes #353, and #418 and #417 - previosuly we misidentified
external lines as being for this file, so the STL includes (for example)
would be misinterpreted as being local. As the STL includes are huge,
we'd try and colourise line 600, or similar.

Tests updated accordingly.