Allow changing url shortening service
This changes so that the name of the module used to do url shortening is
based on an property from configuration.

By default this new 'urlShortenService' property has the value 'none',
which will make it use the 'urlshorten-none.js' module which never
shortens the url. The previous 'urlshorten-google.js' module is
activated by setting it to 'google'.

Hopefully this makes it easy for an standalone compiler-explorer
installation to use an internal url shortening service to avoid leaking
source outside of the organization.