(RFC) Add RISC-V Support
RISC-V is an open standard ISA that's maintained by the non-profit
RISC-V orginization, which has over 100 members today.  More information
can be found at https://riscv.org/ .  Our GCC port is pretty solid as of
7.3.0 -- we recently got more packages built in our Debian port than
IA64 has, for example!

This patch adds support for the RISC-V ISA via GCC 7.3.0 and 8.1.0.
I've tested this via "make check", but I'm afraid that doesn't actually
do anything for this patch so it's not a meaningful test.

We've got a RISC-V setup running already https://cx.rv8.io/, but we'd
love to be part of the official compiler explorer site.

> Note:
> This is more of an RFC than an actual suggested patch, as I can't figure
> out how to add RISC-V to the list of installed cross compilers on the
> official site.  Our GNU toolchain is all upstream, so it should just be
> a matter of adding "riscv64-unknown-linux-gnu" to a list of target
> tuples somewhere.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>