Fix Rust Max_array examples (#1960)
* Fix Max_array.rs iter_mut is not necessary

The previous example failed to compile as `y` was not borrowed as a mutable reference but used to produce a mutable iterator. Changing from `iter_mut()` to `iter()` resolved this issue.

* Fix Max_array_(Optimized).rs

The previous example failed to compile as `y` was not borrowed as a mutable reference but used to produce a mutable iterator. Changing from `iter_mut()` to `iter()` resolved this issue.