Fixes to Armv7 and Armv8 Clang
The previous invocations of these compilers used an x86_64 gcc toolchain,
which prevented Clang from finding suitable headers and libs (eg stdlibc++),
which presented as bugs such as failures to find some C++ headers, and in
incorrect type sizes.

I've also added a sysroot option to the commandline.  This probably isn't
strictly neccessary right now, but it would be necessary if/when execution
support is added.

Change-Id: If6bc3c3d7eb548cd249850ca2992fc28ceb56559