Demangle optimization
Uses a binary search (upper_bound) to find replacements in each
line, instead of looping over all replacements for all lines.
Adds a large test (which was part of bug 1336). On my machine it
used to take 45s to process that test, now it takes ~1.5s.

Hopefully will heavily mitigate #1336.