Check both "9" and "6" length prefixes.
Long story, we used to store everything with a 6 char prefix.
Then we upgraded new links to use 9. That's fine - old 6-char
prefixes were "substring" to their first 9 chars before being
looked up, and that was, still a length six thing, so old links
were found.

However, any "clashing" 6-string, or any "vanity" link (e.g. the
ones _I_ made...), could be longer than 6 chars. They were
originally stored in the DB with prefix 6, but they'd be looked up
with up to 9 chars.

The new links of course are stored with 9 chars of prefix. This is
arguably a mistake, but we're stuck with both 6- and 9- char prefixes
in the database.

For now, we will try looking at the (up to 9) prefix first. If we miss,
we try the 6-prefix (assuming it's different from the up-to-9).

In practice we should _super rarely_ miss, so this shouldn't materially
affect the database access rate.

In future we might want to fix the prefix, and not have it be the same
as the MIN_CHARS thing. And then we should rewrite the entries in
the database.

But for now, this should fix godbolt.org/z/acm19_count4 and similar.

Closes #2582