Initial support for mrustc
Add the needed base for new mrustc compiler.
https://github.com/thepowersgang/mrustc

mrustc is transpiling Rust to C and is mainly used for bootstraping rustc.

This change treats the C output as the final product (as is assembly for
GCC/clang). It is not easily possible to automatically plug this C to other C
compilers to get assembly or executable.

fixes #2643