Create a seperate window for editing compiler flags
If you have a lot of compiler flags, you don't want to be editing
them in the tiny text box. So add a menu option above
'Popular Arguments' that lets you break out the flags into a
separate window, where you can edit them; one-per-line.

When this happens, the regular flags text box becomes disabled,
and the popular arguments dropdown is not populated.