Add generic keywords to OpenCL C and C++ for OpenCL
The generic and __generic keywords are currently reserved in the OpenCL C
specification. They have already been added as keywords in Clang. Without
highlighting, it can seem as if generic is not an address space, as the
other address spaces are highlighted. This can be seen in the AST viewer
on Compiler Explorer. The lack of highlighting makes the AST less convenient
from a user-perspective.

C++ for OpenCL inherits address space behavior from OpenCL C v2.0 s6.5, so
this change applies to both OpenCL C and C++ for OpenCL.