Enable the GCC dump for the rustc using GCC backend
Beware that the way it is working causes several dump files to be empty. The
list of potential files is based on `-fdump-passes` output, but during the
actual compilation, some passes may or may not emit something (eg. a passe may
not be executed at all because it's not applicable). The drop down list will
contain some dump names that will simply display message like:

   Pass 'cselim (tree)' was requested
   but nothing was dumped. Possible causes are:
    - pass is not valid in this (maybe you changed the compiler options);
    - pass is valid but did not emit anything (eg. it was not executed).

This is expected and until someone has a better idea, this will be the case :)

fixes: #2868
Signed-off-by: Marc Poulhiès <dkm@kataplop.net>