Refactor how GCC and GNAT dump are created
Instead of having a specific invocation of the compilers, inject needed options
in the main compilation command. This should fix GNAT race condition on ALI file
creation.

Add optionsForBackend in compiler class that is meant to be overloaded by
compilers for injecting options based on current backend options.

Fixes #3041
References #3029

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>