Fix parsing assembler from binary when templates are used (#3238)
When compiling to binary and with the demangling enabled, the assembler parser
was not correctly matching the "labels". The labelRe regex is tailored from
matching against the real symbol names, before any demangling. In the
binary+demangle context, we rely on objdump doing both at once, feeding our
parser with demangled names.

Make the labelRe match anything of the form:
00000000123af    <every thing here is the label>:

fixes #3235

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>