asm-parser: remove library labels only after reading the whole function
In some cases Dart adds a reference to the source file not right after
a label. This change avoids discarding the instructions between the label
and the first mention of the source file.
Instead, a label is only removed after the whole content of the function did
not contain a reference to the source file.