Ada: Fix inputFileName at incorrect loc for binary and execution
When compiling to binary or running the executable, extra options are inserted,
breaking some weak asumption in Ada/GNAT on where in the list is the
inputFilename.

Override prepareArguments() for GNAT and (more) cleanly handle the
options (split between gnatmake, compiler, binder and linker).

Also use 'gnatmake' as main compiler executable instead of 'gnat'.

Fixes #3709

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>