Add the jakt programming language
See https://github.com/SerenityOS/jakt for the language repository.

Jakt is a new, currently experimental, compiled programming language
designed for the SerenityOS project, especially for OO-style GUI
development. But if the language turns out to be useful, it will
probably be expanded to work for all kinds of systems programming in
SerenityOS.

It's current design for the forseeable future includes a transpilation step
to C++, using parts of the SerenityOS standard library (called AK)
as a runtime to provide underlying implementations for standard types
such as String, HashMap, Array, etc.

As of this point, there are two compiler implementations. The original
implementation is Rust-based, and a selfhosted Jakt compiler is being
developed to replace the Rust-based compiler soon. Then support for the
Rust-based compiler will be dropped.

The Jakt language is still in active development, and many major
features are not yet designed or implemented. At its core is a
philosophy of accurately tracking errors, including allocation failures,
by using value-based exceptions that are part of the function's
signature.

Another strong tenant is first-class C++-interoperability, as most current
SerenityOS code is written in C++.

Due to the included runtime, directly looking at assembly output is
possible but not too insightful. The main point of interest is looking
at the generated C++ code.