Add RGA compiler support to HLSL (#3961)
* Add RGA compiler support to HLSL

AMD's Radeon GPU Analyzer (RGA) tool is an indispensible tool for
graphics programmers to inspect the instructions that will be
compiled by the driver for a given shader and GPU. The integration of
this compiler into CE is slightly unconventional for reasons to be
explained here:

When trying to compile HLSL directly to ISA, RGA requires a full
pipeline state description (specified for graphics pipelines using a
gpso file), and a root signature, since changes to either affect how
resources and data within the shader are accessed or written to.
Specifying both is difficult to the extent that it would add significant
usage friction to the tool. An informal survey among other senior
graphics programming practitioners suggested unanimous agreement to
infer both pipeline state and root signature where possible, with the
expectation that if more accurate ISA code was needed, RGA could be used
directly offline.

Fortunately, RGA supplies an alternative workflow, wherein SPIR-V code
can be compiled directly to the approximate ISA, bypassing both the
pipeline and root signature requirement. To use RGA, the following steps
are performed:

1. Use the default DXC compiler to emit SPIR-V as text to a temporary
   file in the selected temp folder.
2. Compile the ISA using RGA, consuming the output of step 1
3. Rename the resulting file to the output file CE expects

In addition, a non-standard argument --asic is added to the user
options. This argument is filtered for DXC, but for RGA is forwarded as
the selected ASIC to emit ISA for.

These steps are performed by a single `rga.js` script, which is invoked
as an executable. This chaining could have been added within CE library
code directly, but being an atypical flow, felt more appropriate as a
separate script (which has the benefit of faster iteration, due to being
loaded on each compilation request). The paths to DXC and RGA are
supplied as arguments through a simple CLI interface.

NOTE: This commit also adds `-Zi` and `-Qembed_debug` flags to both DXC
and RGA, which provides line association data for DXIL.

Signed-off-by: Jeremy Ong <jeremycong@gmail.com>

* Incorporate PR feedback (see commit body)

This commit introduces several changes:

- HLSL compiler is now Typescript
- A new RGACompiler Typescript class is used to invoke RGA
- The multi-phase compilation that was previously done using a Node
  script now leverages the existing sandboxed execution facilities
- The DXC compiler is configurable as a property on the RGA
  configuration, and an example is provided in hlsl.defaults.properties

As there are several steps needed to compile HLSL to AMD ISA as before,
as steps are done asynchronously so the runtime can continue to service
other requests during compilation (either DXC or RGA) or file I/O.

Signed-off-by: Jeremy Ong <jeremycong@gmail.com>

Signed-off-by: Jeremy Ong <jeremycong@gmail.com>
Co-authored-by: Rubén Rincón Blanco <ruben@rinconblanco.es>