Always include hip/hip_runtime.h
The need to include <hip/hip_runtime.h> and/or <hip/hip_runtime_api.h>
in HIP code before using HIP features is one big difference between
HIP and CUDA. Manually adding it to the command-line here is not ideal,
but the alternative is that the CUDA sample code will be broken by
default when a HIP compiler is selected.

This is perhaps an argument for making HIP a separate language in
compiler explorer, but there's pros and cons to both approaches. I'm
not sure which approach is best, but the existing HIP support that I
built upon was based on using CUDA as the language.