Drop trunk compilers
The nightly builds of ROCm Clang and vanilla Clang lack both device
libraries and the HIP library. Their ability to compile HIP code is
very limited without those, to the extent that they will not be able
to build the example code without changes.

I suspect that it's possible to fix those issues, but it may be less
confusing to drop them from the compiler list until they are fully
functional.