Add ROCm Clang and HIP to CUDA (#4097)
* Add ROCm Clang and HIP to CUDA

* Always include hip/hip_runtime.h

The need to include <hip/hip_runtime.h> and/or <hip/hip_runtime_api.h>
in HIP code before using HIP features is one big difference between
HIP and CUDA. Manually adding it to the command-line here is not ideal,
but the alternative is that the CUDA sample code will be broken by
default when a HIP compiler is selected.

This is perhaps an argument for making HIP a separate language in
compiler explorer, but there's pros and cons to both approaches. I'm
not sure which approach is best, but the existing HIP support that I
built upon was based on using CUDA as the language.

* Add hip-amd to CUDA libs

* Fix id collsion and drop --offload-arch

* Drop trunk compilers

The nightly builds of ROCm Clang and vanilla Clang lack both device
libraries and the HIP library. Their ability to compile HIP code is
very limited without those, to the extent that they will not be able
to build the example code without changes.

I suspect that it's possible to fix those issues, but it may be less
confusing to drop them from the compiler list until they are fully
functional.

* Add self to CONTRIBUTORS.md

* Restore hiptrunk compiler

Put the hiptrunk compiler back how it was, so that the addition of the
HIP libraries is a purely additive change in functionality. However,
mark hiptrunk as hidden so that it is not visible by default.

Co-authored-by: Patrick Quist <partouf@gmail.com>