Change the name of the "auto NSDMI" branch.
The "auto NSDMI" branch has not actually fully supported
non static members in a while, as the patch was initially written
for clang 9, and I have not gotten around to fix it.

The branch now contains a number of experimentals proposals
including
  * parameter pack indexing (https://isocpp.org/files/papers/D2632R0.pdf)
  * magic char8_t cast function
  * _ for variable names
  * non trailing parameter packs
  * constexpr structured bindings
  * constexpr cast from void*
  * And more to come

P2632 is the paper presenting the majority of these features,
and the current plan is to implement as much of it in clang as possible.

The wg21.link link will become active at publication of the paper next week.