Correct gcc-toolchain for "all architectural features" armv8 clang (#4158)
Fixes #4156

Previously these were all set to use an x86_64 toolchain which clearly
isn't right. This lead to things like uintptr_t being smaller than
void*.

Use the same toolchain and sysroot for the "all architectural features"
as for the rest of the configurations.

This also changes the -target to "gnu" not "android" to match the
new settings.

Move all trunk Arm and AArch64 clangs to GCC 12.2.0 sysroot

This includes trunk, trunk with assertions and trunk with
all architectural features.

A newer sysroot is more likely to have the supporting
features needed for new architectural features in these
trunk builds.