Fix llvm-mos clang config file usage.
Manually setting --target=mos on the various llvm-mos targets prevents
the clang driver from picking up the configuration files shipped with
the llvm-mos SDK. This prevents include files from working.

Instead, this change uses -fno-lto -mllvm -zp-avail=224 to set up a
reasonable standalone build configuration. This disables the default LTO
so that assembly can be prduced, and it configures the compiler to
report that the full 256 bytes of zero page are available. (Not
accurate, but good for standalone demonstration purposes.)