Don't try and hack the compilation results (#4429)
* Don't try and hack the compilation results

Carbon tried to synthesize "having run" by overriding and hacking a result routine...but did so _after_ it had been cached. So cached results were broken, but live results not.

This "fixes" by adding a post-compilation, but pre-cache hook and uses that instead. Naming is terrible.

I also took the time to fix the `CompilationResult`'s `buildResult`

Closes #4426