rename: compile to binary -> link to binary
With the merge of #3232, we need to distinguish "compile to binary
object" (compiler + assembler) and "compile to binary" (compiler + assembler +
linker).

Renaming "compile to binary" as "link to binary" is a more correct in this
context and less confusing.

We don't rename the internal field (kept as 'binary') because it's part of the
API and we can't break it.

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>