compile: Synchronise missing non-absolute behaviour with absolute
If compiler.exe is an absolute path that doesn't exist then we print a
warning and return null, resulting in the compiler being disabled and
not shown to the user. However, if compiler.exe is a relative path, as
is the case for a number of compilers in the default configuration, we
instead print an error and continue to create an instance of the
compiler that gets shown to users but does not work. Fix this disparity
and unhelpful behaviour by making the latter do the same as the former.

Fixes https://github.com/compiler-explorer/compiler-explorer/issues/4325