Setup eslint config for lib/ (#4753)
This turns on a number of eslint rules that we setup for static that
aren't specified in the root eslint config as well as ts-eslint rules.
At some point we may want to merge the eslint configs for static/ and
lib/ but we might not be quite ready for that yet. I've temporarily
turned off some rules from static, such as comma-dangle.