Add preliminary support for compiler repositories
Compiler repositories are compiler toolchains which provide access to a
large number of compiler revisions.

A compiler repository is any instance of `ICompiler` which has
`isCompilerLibrary` set to `true`, which means it also implements the
`ICompilerRepository` interface.

If a compiler repository is selected via the UI, a new "revision picker" is
displayed, which provides search and pagination of the available
compiler revisions, in addition to prev/next buttons to browser through
revisions one-by-one.

Pagination is implemented by the `/compilerRevision/:compiler/search`
endpoint, which accepts an optional query, offset and limit.

This change also lays the groundwork for including manyclangs as a
compiler repository.

TODO:

- Implement getRevisions() for manyclangs (via `elfshaker list` or
similar)
- Handle arguments for compiler repository in complie() (we likely
don'ot need to extend the command line as we currently do) and make it
call manyclangs-run or some custom wrapper