Use EXECVE mode instead of ONCE
In ONCE mode, the program is the init process of the pid_namespace which means it won't receive signals for which it hasn't established a signal handler, so if it tries to raise() a signal (e.g. SIGABRT from std::terminate in response to an uncaught exception) that will silently be ignored and the program will behave strangely (e.g. abort() will terminate it with SIGSEGV).

In EXECVE mode, nsjail provides an init process which is parent to the program, so it behaves as expected. Closes #5224 .