Implement comment filtering in LlvmIrParser
This is arguably a bit pointless, since most if not all of the
comments in LLVM IR are filtered out as part of directives or
attributes, but it comes up for things like `opt` on LLVM IR with
debug info and for DXC (which is DXIL, but that's LLVM 3.7-era IR).

In any case, this makes things less confusing in those cases by making
the comments filter actually do something, rather than be ignored.