oneapi 2023.2.1 release + fix sycl execution (#5412)
Depends on: https://github.com/compiler-explorer/infra/pull/1080

oneapi 2023.2.1 release. It is the same compiler as 2023.2.0.

Resolves #5408 
The TBB path was wrong. TBB is used by opencl for CPU, and SYCL
execution for CE target the CPU opencl device
Note that I tested the new configuration on my local setup but did not
reproduce the original problem so it is possible there are other
problems.