Add seastar trunk and 2211 versions to CE
The existing seastar version 1808 is more than five years old and
lots of exciting changes have occurred since then. This patch
adds seastar 2211 from last last year as well as trunk.

The 2211 version of seastar requires PR compiler-explorer/infra#1134
to be accepted first to add it on the infra side.