LLVM TableGen: Add trunk compiler and default include path (#5652)
Since trunk will be a moving target, I've left the default compiler as
17.01.

The standard include path will allow for examples that include files
like `include "llvm/TableGen/SearchableTable.td"`, without extra
options. This is useful if you want to create a minimal example of the
records that an existing backend expects to see.