LLVM TableGen: Add actions to Overrides menu
"actions" are TableGen's version of the target for a compiler
like clang. Where clang might target "Arm" TableGen will target
a backend that produces some format like "gen-searchable-tables".

Actions are gathered from `--help` and appear in the UI as:
```
<option without -->: <description>
```
For example:
```
gen-searchable-tables: Generate generic binary-searchable table
```

The option name means if you are trying replicate an example from
inside llvm, you can match it to the options passed in the build
system. The long description is there if you are just browsing the
options and want to know what it does.

When passed to the compiler I add back the `--` to create the
right option.

The default action is `--print-records` which is the same as
passing no action, which is what I did until now.

With this feature it is possible to make examples like
https://godbolt.org/z/ohK178eMf in a cleaner way than adding
the action option to the compiler options box.