Remove unsupported overrides when the compiler changes
Prior to this, unsupported overrides were kept around as long
as you didn't open the overrides menu. They weren't added to the
compiler's command line, because it would filter them out behind
the scenes, but you could switch back to the old compiler and it'd
be used again.

This changes that so when the compiler changes, overrides that
aren't possible for the new compiler are removed completely.
This means that the Overrides button status and its tooltip
correctly reflect what it's actually using.