Rename opt pipeline without LLVM prefix
This prepares for future work that will reuse the opt pipeline view outside of
the LLVM ecosystem by renaming related components to remove the LLVM prefix.

The pass dumper keep its LLVM prefix, as it is assumed this part is likely to be
customised for each compiler ecosystem.

The historical component name has been preserved in the component list as an
alias to keep old links working.