Remove PreliminaryCompilerInfo creation in d8.ts
app.ts's handlerConfig is made global so that compilers can reference
other existing compilers. This removes the need for faking out a long
config to create a PreliminaryCompilerInfo+compiler for each D8
compilation.

The compiler ID provided to handlerConfig.compileHandler.findCompiler()
must point to a compiler that is installed. In other words, in order to
run a local D8Compiler, the paths for java/javac/kotlinc in their
respective prop files must be valid.

This change also fixes the android-java and android-kotlin defaults file
from '<lang>.default.properties' to '<lang>.defaults.properties'.