Add latest GNU assembler for Arm64
The previous highest version was 2.35.1 and between then and 2.38,
support for many new extensions has been added.

This includes the Scalable Vector Extension (SVE), it's own extension
SVE2, memcopy acceleration (MOPS) and many more.

I checked the version number by installing locally:
```
/opt/compiler-explorer/arm64/gcc-13.2.0/aarch64-unknown-linux-gnu/bin$ ./aarch64-unknown-linux-gnu-as --version
GNU assembler (GNU Binutils) 2.38
```

So that we can handle the new encodings, the objdumper for this
category is also updated to the same version.