Update LLVM Flang file names
The latest builds contain:
```
$ ./bin/flang-
flang-new             flang-to-external-fc
```

The script that was `flang` is now `flang-to-external-fc`.

I've renamed "(old)" to be more specific. I don't think anyone who knew what
"old" meant will be confused by "flang-to-external-fc", and anyone who didn't
know what it meant can at least Google that name and find out that it's a script.