fortran: Enable running binaries with flang-new
This was already done for the default config, this does the same
for the amazon config.

Flang used to require that libpgmath be installed
(https://github.com/flang-compiler/flang/blob/master/docs/libpgmath.rst).

This is no longer the case (https://reviews.llvm.org/D140236),
so we should be able to build executable binaries without it.