flang: Enable llvm-ir viewer for flang-new
flang supports the -emit-llvm option but the difference from
clang is that while clang allows -Xclang -emit-llvm, flang does not.

This is a deliberate choice by the flang driver to make sure the user
isn't asking the compiler to do 2 things at once. In this case assemble
and emit llvm IR as well.

When you -Xclang -emit-llvm, it uses the -o value for the assembly
output and changes the input file extension to .ll for the IR output.

Since we're doing `flang-new ... -emit-llvm`, flang-new will just
output the IR to the output file name we give it.

To do this I've copied code from ISPCCompiler which does the same thing.
We could dedupe this, but I'm not sure whether it's worth it while
we have only 2 compilers doing this.

(clang supports `clang ... -emit-llvm` as well, so perhaps the flang
style could become the majority mode eventually)

Final note, flang-to-external-fc could support this with changes but
users are better of just using flang-new, the result is the same.
(as flang-to-external-fc doesn't advertise -emit-llvm, it won't be
selectable anyway)