Minor improvements to noscript for mobile users  (#6071)
The `input` field for compiler arguments in noscript has `autocomplete`
set to off, which results in the field being cleared between each
compilation on Safari & Samsung Internet.

Also, autocorrect and auto capitalization are enabled in the source code
`textarea`, which makes it incredibly tedious to type anything on
mobile. This PR addresses both of these issues.