Change getCompilerResultLanguageId prototype
The getCompilerResultLanguageId() method can be overridden by compiler
classes to report what mode the monaco editor should be using to display
what the compiler is emitting. The method was provided with the
"filters" (badly named...) object that contains the user selected option
for "link to binary", "compile to object", etc.
Some compilers may need different monaco settings depending on how they
are called.

refs #6079

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>