Disable sharing clipboard button when link generation fails
Fixes #2545

The text box containing the link was already read only. So I
think the intention was always to set the clipboard button to
disabled, but the author used the wrong object by mistake.

I've changed it so that it finds the clipboard button and
disables it on error, and enables it on success.

The modal is constructed once and updated if re-opened,
so we need to set the state each time in case the result has changed.

I tested this locally by making the code always take the
error path.